implement node-style module
